+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 2019-05-17 v1.4.0-SNAPSHOT
  * Added description for data-import flow
- * Fixed issue with saving errorRecords
+ * Fixed issue with saving ErrorRecords
+ * Added batch update of ParsedRecords after assigning Instance id to MARC records
  
 ## 2019-05-17 v1.3.1
  * Filled complete date and stub HrID for Job Execution